Bloqueados, gamificacion

+++ BoostARuaal/app/src/main/java/com/example/boostar_uaal/core/entities/ChallengeSteps.kt
@@ -5,5 +5,5 @@ data class ChallengeStep(
     val multimedia: Int,
     val title: String? = null,
     val text: String,
-    val sleepTimeInMilliseconds: Int
+    val sleepTimeInMilliseconds: Long
 )
